refactor(egovla): move deployment/egovla → examples/realRobots/EgoVLA/eval_files
Consolidate the EgoVLA G1 deployment code under the EgoVLA example folder, as
requested in review, so it sits next to train_files/ (mirrors the
RoboChallenge_table30v2/{train,eval}_files package layout).

- git-mv the server + policy + kinematics + assets into eval_files/ (history
  preserved); add EgoVLA/__init__.py, keep eval_files/__init__.py, so the
  example is a proper package.
- drop the ad-hoc sys.path.insert(..., "..", "..") shims: the repo is an
  editable install, so deployment.model_server / starVLA resolve as installed
  packages (matching the WholeBody eval_files convention). The sibling import in
  server_egovla_g1.py becomes
  `from examples.realRobots.EgoVLA.eval_files.egovla_g1_policy import EgoVLAG1Policy`.
- update README + docstring paths; license-gated *.pth stay git-ignored via
  eval_files/assets/.gitignore.

No behavior change: py_compile + import of all servers pass; no remaining
deployment/egovla references in the tree.
