+++ examples/realRobots/EgoVLA/eval_files/assets/g1_29dof_with_hand.urdf
