+++ examples/realRobots/EgoVLA/eval_files/egovla_g1_policy.py
