+++ examples/realRobots/EgoVLA/eval_files/g1_camera.py
